Port to ATtiny4313 and update interpolation code
The newer chips behaved slightly erratically, the deciseconds sometimes ticked over twice, then back. I think the chip in the original clock, running on its internal oscillator, may have been slightly slow. The newer chips may be slightly fast, so the interpolation counts beyond the .99 mark right before the pulse comes in to roll over. The OSCCAL register could be used to adjust this, but this is only one byte with non-linear scaling. Instead the timer period for the centiseconds can be adjusted. When the 1PPS pulse arrives, we check if the interpolation is running fast or slow, then adjust accordingly. Potentially, the number could be stored in the eeprom once it's hit the mark.