Set timer prescaler to /8
Annoyed I didn't spot this sooner, with prescaler /64 OCR1A was starting at 1250, with /8 we can start at 10000, meaning we can track the PPS interval with higher resolution. This does mean that the initial synchronization takes longer because we're still doing only incremental adjustments, but the improvements to offline accuracy should be worth it.