Add osccal calibration routine
Use stable 32kHz square wave from DS3231SN to calibrate oscillator and write it to EEPROM. Use `make osccal` to run the routine and set the EESave fuse. It takes a few seconds to run. Main source code is updated to make use of the EEPROM calibration if (and only if) it's present.